requirements.txt
mysql-connector-python
psycopg2-binary
apache-beam